dynameat-api/pyproject.toml
[tool.poetry]
name = "dynameat-api"
version = "0.1.0"
description = ""
authors = ["Miguel Ocaña <miguelocanaacosta@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.3"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
pandas = "^2.2.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
